backend/requirements.txt
# Django and Core Dependencies
Django==5.0.2
djangorestframework==3.14.0
django-cors-headers==4.3.1
django-filter==23.5
django-ckeditor==6.7.0
django-storages==1.14.2
django-extensions==3.2.3
django-debug-toolbar==4.2.0

# Database
psycopg2-binary==2.9.9

# Image Processing
Pillow==10.1.0

# Environment Variables
python-decouple==3.8

# AWS (for file storage)
boto3==1.34.0

# Background Tasks
celery==5.3.4
redis==5.0.1

# Development Tools
setuptools==69.0.3
wheel==0.42.0
